Simplify stuff
I am working on the assumption that the frontend will be
written mostly in F#/Fable/Feliz, and hence we can use
F# types to our hearts content. The parts that are in plain
JS can treat them mostly as opaque values.

I extracted the shared types between the backend code
and the frontend code into a single shared file which
doesn't have any dependencies on the reference implementation.
As a result a lot less JS is compiled twice.

Also added in a dummy compile step for when you run `yarn dev`
to avoid failure to start when booting it up.