Move from Snowpack to Webpack and electron-forge
The electron-snowpack plugin wasn't quite up to the task.
For once it would not allow me to build releases at all.

electron-forge which is now used seems to be the more standard
way of doing things, and additionally it's much easier to
understand what is going on and what needs to be configured where.