Prevent Electron from packing and unpacking our data
During IPC Electron encodes everything and decodes it
as JSON. This is slow, which we could live with, but
fails on large payloads. This is noticeable in that
the anonymized results of large datasets (say a 30mb CSV),
fails to materialize in the frontend and we are given
a error that we have blown the stack.

This commit moves the serialization into our own code.
It does it in a slightly clumsy way that intends to ensure
that no smarts will try to decode it for us.